LAW_FIRM/ Added SubResources functionality

+++ src/main/java/com/mogilan/servlet/LawyerServlet.java
@@ -3,11 +3,15 @@ package com.mogilan.servlet;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import com.mogilan.exception.PathVariableException;
 import com.mogilan.exception.handler.ServletExceptionHandler;
 import com.mogilan.exception.handler.impl.ServletExceptionHandlerImpl;
 import com.mogilan.service.LawyerService;
+import com.mogilan.service.TaskService;
 import com.mogilan.service.impl.LawyerServiceImpl;
+import com.mogilan.service.impl.TaskServiceImpl;
 import com.mogilan.servlet.dto.LawyerDto;
+import com.mogilan.servlet.dto.TaskDto;
 import com.mogilan.util.ServletsUtil;
 import jakarta.servlet.ServletConfig;
 import jakarta.servlet.ServletException;
@@ -18,15 +22,19 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @WebServlet("/api/lawyers/*")
 public class LawyerServlet extends HttpServlet {
 
     private final LawyerService lawyerService = LawyerServiceImpl.getInstance();
+    private final TaskService taskService = TaskServiceImpl.getInstance();
     private final ServletExceptionHandler exceptionHandler = ServletExceptionHandlerImpl.getInstance();
     private final ObjectMapper objectMapper = new ObjectMapper();
-    private static final String UPDATED_MESSAGE = "UPDATED";
-    private static final String DELETED_MESSAGE = "DELETED";
+
+    private static final String REGEX_FOR_SUB_RESOURCES_1 = "\\/(\\d+)\\/tasks\\/";
+    private static final String REGEX_FOR_SUB_RESOURCES_2 = "\\/(\\d+)\\/tasks\\/(\\d*)";
 
     @Override
     public void init(ServletConfig config) throws ServletException {
@@ -39,7 +47,11 @@ public class LawyerServlet extends HttpServlet {
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) {
         try {
             var pathInfo = req.getPathInfo();
-            if (pathInfo == null || pathInfo.equals("/")) {
+            if (pathInfo != null &&
+                    (Pattern.matches(REGEX_FOR_SUB_RESOURCES_1, pathInfo) ||
+                            Pattern.matches(REGEX_FOR_SUB_RESOURCES_2, pathInfo))) {
+                processDoGetForSubResources(req, resp);
+            } else if (pathInfo == null || pathInfo.equals("/")) {
                 var lawyerDtoList = lawyerService.readAll();
                 if (!lawyerDtoList.isEmpty()) {
                     writeJsonResponse(resp, HttpServletResponse.SC_OK, lawyerDtoList);
@@ -59,9 +71,18 @@ public class LawyerServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) {
         try {
-            var lawyerDto = readRequestJson(req, LawyerDto.class);
-            var createdLawyerDto = lawyerService.create(lawyerDto);
-            writeJsonResponse(resp, HttpServletResponse.SC_CREATED, createdLawyerDto);
+            var pathInfo = req.getPathInfo();
+            if (pathInfo != null &&
+                    (Pattern.matches(REGEX_FOR_SUB_RESOURCES_1, pathInfo) ||
+                            Pattern.matches(REGEX_FOR_SUB_RESOURCES_2, pathInfo))) {
+                processDoPostForSubResources(req, resp);
+            } else if (pathInfo != null && !pathInfo.equals("/")) {
+                throw new PathVariableException("Method PUT doesn't support path variable");
+            } else {
+                var lawyerDto = readRequestJson(req, LawyerDto.class);
+                var createdLawyerDto = lawyerService.create(lawyerDto);
+                writeJsonResponse(resp, HttpServletResponse.SC_CREATED, createdLawyerDto);
+            }
         } catch (Exception e) {
             exceptionHandler.handleException(resp, e);
         }
@@ -71,10 +92,16 @@ public class LawyerServlet extends HttpServlet {
     protected void doPut(HttpServletRequest req, HttpServletResponse resp) {
         try {
             var pathInfo = req.getPathInfo();
-            var id = ServletsUtil.getPathVariable(pathInfo);
-            var lawyerDto = readRequestJson(req, LawyerDto.class);
-            lawyerService.update(id, lawyerDto);
-            writeJsonResponse(resp, HttpServletResponse.SC_OK, UPDATED_MESSAGE);
+            if (pathInfo != null &&
+                    (Pattern.matches(REGEX_FOR_SUB_RESOURCES_1, pathInfo) ||
+                            Pattern.matches(REGEX_FOR_SUB_RESOURCES_2, pathInfo))) {
+                processDoPutForSubResources(req, resp);
+            } else {
+                var id = ServletsUtil.getPathVariable(pathInfo);
+                var lawyerDto = readRequestJson(req, LawyerDto.class);
+                lawyerService.update(id, lawyerDto);
+                writeJsonResponse(resp, HttpServletResponse.SC_OK, ServletsUtil.UPDATED_MESSAGE);
+            }
         } catch (Exception e) {
             exceptionHandler.handleException(resp, e);
         }
@@ -84,11 +111,86 @@ public class LawyerServlet extends HttpServlet {
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) {
         try {
             var pathInfo = req.getPathInfo();
-            var id = ServletsUtil.getPathVariable(pathInfo);
-            lawyerService.deleteById(id);
-            writeJsonResponse(resp, HttpServletResponse.SC_OK, DELETED_MESSAGE);
+            if (pathInfo != null &&
+                    (Pattern.matches(REGEX_FOR_SUB_RESOURCES_1, pathInfo) ||
+                            Pattern.matches(REGEX_FOR_SUB_RESOURCES_2, pathInfo))) {
+                processDoDeleteForSubResources(req, resp);
+            } else {
+                var id = ServletsUtil.getPathVariable(pathInfo);
+                lawyerService.deleteById(id);
+                writeJsonResponse(resp, HttpServletResponse.SC_OK, ServletsUtil.DELETED_MESSAGE);
+            }
         } catch (Exception e) {
             exceptionHandler.handleException(resp, e);
+        }
+    }
+
+    private void processDoGetForSubResources(HttpServletRequest req, HttpServletResponse resp) throws IOException, PathVariableException {
+        var ids = getResourcesIds(req);
+        var lawyerId = ids.resourceId();
+        var taskId = ids.subResourceId();
+        if (lawyerId != null && taskId == null) {
+            var taskDtoList = taskService.readAllByLawyerId(lawyerId);
+            if (!taskDtoList.isEmpty()) {
+                writeJsonResponse(resp, HttpServletResponse.SC_OK, taskDtoList);
+            } else {
+                writeJsonResponse(resp, HttpServletResponse.SC_NO_CONTENT, taskDtoList);
+            }
+        } else if (lawyerId != null) {
+            if (taskService.isLawyerResponsibleForTask(taskId, lawyerId)) {
+                var taskDto = taskService.readById(taskId);
+                writeJsonResponse(resp, HttpServletResponse.SC_OK, taskDto);
+            } else {
+                writeJsonResponse(resp, HttpServletResponse.SC_NOT_FOUND, ServletsUtil.NOT_FOUND_MESSAGE);
+            }
+        } else {
+            throw new PathVariableException(ServletsUtil.PATH_VARIABLE_IS_NOT_CORRECT_MESSAGE);
+        }
+    }
+
+    private void processDoPostForSubResources(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+        var ids = getResourcesIds(req);
+        var lawyerId = ids.resourceId();
+        var taskId = ids.subResourceId();
+        if (lawyerId != null && taskId == null) {
+            var taskDto = readRequestJson(req, TaskDto.class);
+            var createdTaskDto = taskService.create(taskDto);
+            writeJsonResponse(resp, HttpServletResponse.SC_CREATED, createdTaskDto);
+        } else {
+            throw new PathVariableException(ServletsUtil.PATH_VARIABLE_IS_NOT_CORRECT_MESSAGE);
+        }
+    }
+
+    private void processDoPutForSubResources(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+        var ids = getResourcesIds(req);
+        var lawyerId = ids.resourceId();
+        var taskId = ids.subResourceId();
+        if (taskId != null && lawyerId != null) {
+            if (taskService.isLawyerResponsibleForTask(taskId, lawyerId)) {
+                var taskDto = readRequestJson(req, TaskDto.class);
+                taskService.update(lawyerId, taskDto);
+                writeJsonResponse(resp, HttpServletResponse.SC_OK, ServletsUtil.UPDATED_MESSAGE);
+            } else {
+                writeJsonResponse(resp, HttpServletResponse.SC_BAD_REQUEST, ServletsUtil.BAD_REQUEST_MESSAGE);
+            }
+        } else {
+            throw new PathVariableException(ServletsUtil.PATH_VARIABLE_IS_NOT_CORRECT_MESSAGE);
+        }
+    }
+
+    private void processDoDeleteForSubResources(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+        var ids = getResourcesIds(req);
+        var lawyerId  = ids.resourceId();
+        var taskId = ids.subResourceId();
+        if (taskId != null && lawyerId != null) {
+            if(taskService.isLawyerResponsibleForTask(taskId, lawyerId)){
+                taskService.deleteById(taskId);
+                writeJsonResponse(resp, HttpServletResponse.SC_OK, ServletsUtil.DELETED_MESSAGE);
+            } else {
+                writeJsonResponse(resp, HttpServletResponse.SC_BAD_REQUEST, ServletsUtil.BAD_REQUEST_MESSAGE);
+            }
+        } else {
+            throw new PathVariableException(ServletsUtil.PATH_VARIABLE_IS_NOT_CORRECT_MESSAGE);
         }
     }
 
@@ -107,5 +209,29 @@ public class LawyerServlet extends HttpServlet {
         try (var writer = resp.getWriter()) {
             objectMapper.writeValue(writer, dto);
         }
+    }
+
+    private TaskServlet.Ids getResourcesIds(HttpServletRequest req) throws PathVariableException {
+        var pathInfo = req.getPathInfo();
+        Pattern pattern = Pattern.compile(REGEX_FOR_SUB_RESOURCES_2);
+        Matcher matcher = pattern.matcher(pathInfo);
+        Long resourceId = null;
+        Long subResourceId = null;
+        if (matcher.find()) {
+            var group1 = matcher.group(1);
+            if (!group1.isBlank()) {
+                resourceId = Long.parseLong(group1);
+            }
+            var group2 = matcher.group(2);
+            if (!group2.isBlank()) {
+                subResourceId = Long.parseLong(group2);
+            }
+        } else {
+            throw new PathVariableException(ServletsUtil.PATH_VARIABLE_IS_NOT_CORRECT_MESSAGE);
+        }
+        return new TaskServlet.Ids(resourceId, subResourceId);
+    }
+
+    record Ids(Long resourceId, Long subResourceId) {
     }
 }
